+++ src/main/java/mitrack_proyecto/mitrack/model/Usuario.java
@@ -12,15 +12,20 @@ import jakarta.persistence.ManyToOne;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 
+//Genera automaticamente getters y setters toStringetc() ahorro.
 @Data
+// Genera un constructor con todos los campos como parametro
 @AllArgsConstructor
+// Genera un constructor vacio (sin argumentos)
 @NoArgsConstructor
-
 public class Usuario {
 
+    // Define clave primaria
+    // Se indica que esta id se generara de manera automatica de forma
+    // auto-incremental
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id_usuario;
+    private Long id_usuario;
     private String nombre;
     @Email
     @NotBlank
@@ -28,6 +33,6 @@ public class Usuario {
     @NotBlank
     private String contraseña;
     @ManyToOne(fetch = FetchType.EAGER)
-    @JoinColumn(name = "id_rol") // nombre columna FK en tabla usuario
+    @JoinColumn(name = "id_rol")
     private Rol rol;
 }
